Subsitute all JSONB models with its implementations (and move config file to lib)
