+++ shogun-boot/src/main/java/de/terrestris/shogun/boot/config/BootSwaggerConfig.java
@@ -16,7 +16,7 @@
  */
 package de.terrestris.shogun.boot.config;
 
-import de.terrestris.shogun.config.SwaggerConfig;
+import de.terrestris.shogun.lib.config.SwaggerConfig;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.annotation.Configuration;
 import springfox.documentation.builders.PathSelectors;
